I have "hardened" the code in blendheaders.remove_distortion_keywords so that it does not crash when 'del hdr[...]' is called. The original cause of the crash is a bug in PyFITS 3.2.1 (header.__del__).
NOTE: This is a *temporary* workaround the PyFITS bug and should be removed once the bug is fixed.

To reproduce the PyFITS error on a RAW ACS file, follow these steps:

Try to delete 'DP1.EXTVER*' first. It works. Any repeated attempt to delete the same word works too. However, it is when one tries to delete a different item that PyFITS crashes. So,

--> import pyfits
--> pyfits.__version__
'3.2.1'
--> hdu=pyfits.open('j8eh11j5q_raw.fits')
--> hdr=hdu[1].header
--> del hdr['DP1.EXTVER*']
--> del hdr['DP1.NAXES*']
Traceback (innermost last):
  File "<console>", line 1, in <module>
  File "/Users/mcara/STScI/opus1.3/Ureka/variants/common/lib/python2.7/site-packages/pyfits/header.py", line 206, in __delitem__
    del self[idx]
  File "/Users/mcara/STScI/opus1.3/Ureka/variants/common/lib/python2.7/site-packages/pyfits/header.py", line 248, in __delitem__
    indices.remove(idx)
ValueError: list.remove(x): x not in list


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/fitsblender/trunk@30551 fe389314-cf27-0410-b35b-8c050e845b92